meet-backend/requirements.txt
pyjwt==2.0.0
psycopg2